rename and export from index
