+++ packages/server/utils/__tests__/rateLimiters/InMemoryRateLimiter.test.ts
@@ -1,5 +1,8 @@
 import ms from 'ms'
-import {InMemoryRateLimiter, InMemoryRateLimiterConfig} from '../../rateLimits/InMemoryRateLimiter'
+import {
+  InMemoryRateLimiter,
+  InMemoryRateLimiterConfig
+} from '../../rateLimiters/InMemoryRateLimiter'
 
 describe('InMemoryRateLimiter', () => {
   function subject(configOverrides: Partial<InMemoryRateLimiterConfig> = {}): InMemoryRateLimiter {
